busy indicator, tooltips
Busy indicator and tooltips associated with mirrors were added. All the
features should work both online and offline (package/desktop version).
Tooltips are displayed whenever a user hovers over the out-link text
"here" in footer. The busy indicator GIF is free available, so that we
can use it with no restrictions; it is displayed whenever the
applications is performing some calculations and is "busy".