change of report downloading logic
If you will agree, I'd suggest a change of logic of the report
downloading: firstly, a user clicks on generate button (so far so good),
but, the download button is not available at this moment. When the user
starts generating the report, message box with some time consumption
warnings will pop up (as expected), then the process of report
generating will follow, while progress bar is displayed (as expected as
well).  The download button is still not available up to now and is
dynamically rendered just at the moment the report is prepared.