padding-top added to mobile version
I added padding-top settings applied on mobile screens. From now, there
should be a sufficient space between the header and the body text
whenever is the app displayed on mobile device.

(However, there is a little bit unavoidable feature (but not bug): when
a screen width is about 1000 px, navbars in the header are aligned below
themselves and overlap some body headings and text. It could be managed
by some HTML and bootstrap.css changes.)