Minor changes -- opacity of the footer for textareas like "No file selected"
The footer of the application is transparent to placeholders of
text-areas such as "No file selected" placed on the second tab "Data",
i. e. the placeholders like "No file selected" are displayed in front of
the footer when scrolled at the same level. I suppose this is
undesirable, therefore I have increased an opacity of the footer to its
maximum just to avoid the accidental appearance of the text-areas in
front of the footer.